moved tray addons manager to tray
