+++ client/ayon_core/modules/base.py
@@ -3,7 +3,6 @@
 from ayon_core.addon import (
     AYONAddon,
     AddonsManager,
-    TrayAddonsManager,
     load_addons,
 )
 from ayon_core.addon.base import (
@@ -12,18 +11,15 @@ from ayon_core.addon.base import (
 )
 
 ModulesManager = AddonsManager
-TrayModulesManager = TrayAddonsManager
 load_modules = load_addons
 
 
 __all__ = (
     "AYONAddon",
     "AddonsManager",
-    "TrayAddonsManager",
     "load_addons",
     "OpenPypeModule",
     "OpenPypeAddOn",
     "ModulesManager",
-    "TrayModulesManager",
     "load_modules",
 )
